refactoring PodTemplate
